Clear current retro and archive
Allow user to create a new retro by archiving the previous retro and
clearing the screen.

Allow user to clear the screen.

When clearing the screen it maintiains prevous headers because it is
assumed that the format will be similar. Using the `persist` tag on a
headline will leave that heading and its content untouched.